adding option --invert-sg-removal
added the option for --invert-sg-removal and teh enccessary code to do this, when this option is defined the list of SG(s) specified - comma separated in teh command line options for --g will be the only SG(sg) left attached to teh resource.

all the appropriate checking to ensure there is always one sg left on the instance in the fucntion.

Slection logic for wether the option is on or off has been implemeted.

aligned the options so skip did the right thing